requirements.txt
requests>=2.31.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
webcolors>=1.13.0
python-dotenv>=1.0.0